Add CoreAirplaneMovement test module
Move AirplaneMovement tests to a Core-level test module so they directly
test CoreAirplaneMovement behavior. Keep a slim AirplaneMovement test
module that verifies the public subclass relationship, the stricter type
check on wing_movements, and inherited behavior. Duplicate the three
AirplaneMovement fixtures that consumer tests need since Movement
enforces the stricter public type.
