+++ tests/unit/__init__.py
@@ -18,6 +18,9 @@ This package contains the following modules:
 
     test_airplane_movement.py: This module contains classes to test AirplaneMovements.
 
+    test_core_airplane_movement.py: This module contains classes to test
+    CoreAirplaneMovements.
+
     test_core.py: This module contains classes to test the core classes in _core.py.
 
     test_core_wing_cross_section_movement.py: This module contains classes to test
@@ -77,6 +80,7 @@ import tests.unit.test_airfoil
 import tests.unit.test_airplane
 import tests.unit.test_airplane_movement
 import tests.unit.test_core
+import tests.unit.test_core_airplane_movement
 import tests.unit.test_core_wing_cross_section_movement
 import tests.unit.test_core_wing_movement
 import tests.unit.test_horseshoe_vortex
