+++ tests/unit/test_airplane_movement.py
@@ -1,11 +1,6 @@
 """This module contains classes to test AirplaneMovements."""
 
-import copy
 import unittest
-
-import numpy as np
-import numpy.testing as npt
-from scipy import signal
 
 import pterasoftware as ps
 from tests.unit.fixtures import (
@@ -18,1208 +13,56 @@ from tests.unit.fixtures import (
 class TestAirplaneMovement(unittest.TestCase):
     """This is a class with functions to test AirplaneMovements."""
 
-    @classmethod
-    def setUpClass(cls):
-        """Set up test fixtures once for all AirplaneMovement tests."""
-        # Spacing test fixtures.
-        cls.sine_spacing_Cg_airplane_movement = (
-            airplane_movement_fixtures.make_sine_spacing_Cg_airplane_movement_fixture()
-        )
-        cls.uniform_spacing_Cg_airplane_movement = (
-            airplane_movement_fixtures.make_uniform_spacing_Cg_airplane_movement_fixture()
-        )
-        cls.mixed_spacing_Cg_airplane_movement = (
-            airplane_movement_fixtures.make_mixed_spacing_Cg_airplane_movement_fixture()
-        )
-
-        # Additional test fixtures.
-        cls.static_airplane_movement = (
-            airplane_movement_fixtures.make_static_airplane_movement_fixture()
-        )
-        cls.basic_airplane_movement = (
-            airplane_movement_fixtures.make_basic_airplane_movement_fixture()
-        )
-        cls.Cg_airplane_movement = (
-            airplane_movement_fixtures.make_Cg_airplane_movement_fixture()
-        )
-        cls.phase_offset_Cg_airplane_movement = (
-            airplane_movement_fixtures.make_phase_offset_Cg_airplane_movement_fixture()
-        )
-        cls.multiple_periods_airplane_movement = (
-            airplane_movement_fixtures.make_multiple_periods_airplane_movement_fixture()
-        )
-        cls.custom_spacing_Cg_airplane_movement = (
-            airplane_movement_fixtures.make_custom_spacing_Cg_airplane_movement_fixture()
-        )
-        cls.mixed_custom_and_standard_spacing_airplane_movement = (
-            airplane_movement_fixtures.make_mixed_custom_and_standard_spacing_airplane_movement_fixture()
-        )
-
-    def test_spacing_sine_for_Cg_GP1_CgP1(self):
-        """Test that sine spacing actually produces sinusoidal motion for Cg_GP1_CgP1."""
-        num_steps = 10
-        delta_time = 0.01
-        airplanes = self.sine_spacing_Cg_airplane_movement.generate_airplanes(
-            num_steps=num_steps,
-            delta_time=delta_time,
-        )
-
-        # Extract positions (in Earth axes, relative to the simulation starting
-        # point) from generated Airplanes.
-        x_positions = np.array([airplane.Cg_GP1_CgP1[0] for airplane in airplanes])
-
-        # Calculate expected sine wave values.
-        times = np.linspace(0, num_steps * delta_time, num_steps, endpoint=False)
-        expected_x = 0.1 * np.sin(2 * np.pi * times / 1.0)
-
-        # Assert that the generated positions match the expected sine wave.
-        npt.assert_allclose(x_positions, expected_x, rtol=1e-10, atol=1e-14)
-
-    def test_spacing_uniform_for_Cg_GP1_CgP1(self):
-        """Test that uniform spacing actually produces triangular wave motion for
-        Cg_GP1_CgP1."""
-        num_steps = 10
-        delta_time = 0.01
-        airplanes = self.uniform_spacing_Cg_airplane_movement.generate_airplanes(
-            num_steps=num_steps,
-            delta_time=delta_time,
-        )
-
-        # Extract positions (in Earth axes, relative to the simulation starting
-        # point) from generated Airplanes.
-        x_positions = np.array([airplane.Cg_GP1_CgP1[0] for airplane in airplanes])
-
-        # Calculate expected triangular wave values.
-        times = np.linspace(0, num_steps * delta_time, num_steps, endpoint=False)
-        expected_x = 0.1 * signal.sawtooth(2 * np.pi * times / 1.0 + np.pi / 2, 0.5)
-
-        # Assert that the generated positions match the expected triangular wave.
-        npt.assert_allclose(x_positions, expected_x, rtol=1e-10, atol=1e-14)
-
-    def test_spacing_mixed_for_Cg_GP1_CgP1(self):
-        """Test that mixed spacing types work correctly for Cg_GP1_CgP1."""
-        num_steps = 10
-        delta_time = 0.01
-        airplanes = self.mixed_spacing_Cg_airplane_movement.generate_airplanes(
-            num_steps=num_steps,
-            delta_time=delta_time,
-        )
-
-        # Extract positions (in Earth axes, relative to the simulation starting
-        # point) from generated Airplanes.
-        x_positions = np.array([airplane.Cg_GP1_CgP1[0] for airplane in airplanes])
-        y_positions = np.array([airplane.Cg_GP1_CgP1[1] for airplane in airplanes])
-        z_positions = np.array([airplane.Cg_GP1_CgP1[2] for airplane in airplanes])
-
-        # Calculate expected values.
-        times = np.linspace(0, num_steps * delta_time, num_steps, endpoint=False)
-        expected_x = 0.1 * np.sin(2 * np.pi * times / 1.0)
-        expected_y = 0.08 * signal.sawtooth(2 * np.pi * times / 1.0 + np.pi / 2, 0.5)
-        expected_z = 0.06 * np.sin(2 * np.pi * times / 1.0)
-
-        # Assert that the generated positions match the expected values.
-        npt.assert_allclose(x_positions, expected_x, rtol=1e-10, atol=1e-14)
-        npt.assert_allclose(y_positions, expected_y, rtol=1e-10, atol=1e-14)
-        npt.assert_allclose(z_positions, expected_z, rtol=1e-10, atol=1e-14)
-
-    def test_base_airplane_validation(self):
-        """Test that base_airplane parameter validation works correctly."""
-        # Test non-Airplane raises error.
-        with self.assertRaises(TypeError):
-            ps.movements.airplane_movement.AirplaneMovement(
-                base_airplane="not an airplane",
-                wing_movements=[
-                    wing_movement_fixtures.make_static_wing_movement_fixture()
-                ],
+    def test_is_subclass_of_core(self):
+        """Test that AirplaneMovement is a subclass of CoreAirplaneMovement."""
+        self.assertTrue(
+            issubclass(
+                ps.movements.airplane_movement.AirplaneMovement,
+                ps._core.CoreAirplaneMovement,
             )
+        )
 
-        # Test None raises error.
-        with self.assertRaises(TypeError):
-            ps.movements.airplane_movement.AirplaneMovement(
-                base_airplane=None,
-                wing_movements=[
-                    wing_movement_fixtures.make_static_wing_movement_fixture()
-                ],
-            )
-
-        # Test valid Airplane works.
+    def test_instantiation_returns_correct_type(self):
+        """Test that AirplaneMovement instantiation returns an AirplaneMovement."""
         base_airplane = geometry_fixtures.make_first_airplane_fixture()
         wing_movements = [wing_movement_fixtures.make_static_wing_movement_fixture()]
-        airplane_movement = ps.movements.airplane_movement.AirplaneMovement(
-            base_airplane=base_airplane, wing_movements=wing_movements
-        )
-        self.assertEqual(airplane_movement.base_airplane, base_airplane)
-
-    def test_ampCg_GP1_CgP1_validation(self):
-        """Test ampCg_GP1_CgP1 parameter validation."""
-        base_airplane = geometry_fixtures.make_first_airplane_fixture()
-        wing_movements = [wing_movement_fixtures.make_static_wing_movement_fixture()]
-
-        # Test valid values.
-        valid_amps = [
-            (0.0, 0.0, 0.0),
-            (1.0, 2.0, 3.0),
-            [0.5, 1.5, 2.5],
-            np.array([0.1, 0.2, 0.3]),
-        ]
-        for amp in valid_amps:
-            with self.subTest(amp=amp):
-                airplane_movement = ps.movements.airplane_movement.AirplaneMovement(
-                    base_airplane=base_airplane,
-                    wing_movements=wing_movements,
-                    ampCg_GP1_CgP1=amp,
-                )
-                npt.assert_array_equal(airplane_movement.ampCg_GP1_CgP1, amp)
-
-        # Test negative values raise error.
-        with self.assertRaises(ValueError):
-            ps.movements.airplane_movement.AirplaneMovement(
-                base_airplane=base_airplane,
-                wing_movements=wing_movements,
-                ampCg_GP1_CgP1=(-1.0, 0.0, 0.0),
-            )
-
-        # Test invalid types raise error.
-        # noinspection PyTypeChecker
-        with self.assertRaises((TypeError, ValueError)):
-            ps.movements.airplane_movement.AirplaneMovement(
-                base_airplane=base_airplane,
-                wing_movements=wing_movements,
-                ampCg_GP1_CgP1="invalid",
-            )
-
-    def test_periodCg_GP1_CgP1_validation(self):
-        """Test periodCg_GP1_CgP1 parameter validation."""
-        base_airplane = geometry_fixtures.make_first_airplane_fixture()
-        wing_movements = [wing_movement_fixtures.make_static_wing_movement_fixture()]
-
-        # Test valid values.
-        valid_periods = [(0.0, 0.0, 0.0), (1.0, 2.0, 3.0), [0.5, 1.5, 2.5]]
-        for period in valid_periods:
-            with self.subTest(period=period):
-                # Need matching amps for non-zero periods.
-                amp = tuple(1.0 if p > 0 else 0.0 for p in period)
-                airplane_movement = ps.movements.airplane_movement.AirplaneMovement(
-                    base_airplane=base_airplane,
-                    wing_movements=wing_movements,
-                    ampCg_GP1_CgP1=amp,
-                    periodCg_GP1_CgP1=period,
-                )
-                npt.assert_array_equal(airplane_movement.periodCg_GP1_CgP1, period)
-
-        # Test negative values raise error.
-        with self.assertRaises(ValueError):
-            ps.movements.airplane_movement.AirplaneMovement(
-                base_airplane=base_airplane,
-                wing_movements=wing_movements,
-                ampCg_GP1_CgP1=(1.0, 1.0, 1.0),
-                periodCg_GP1_CgP1=(-1.0, 1.0, 1.0),
-            )
-
-    def test_spacingCg_GP1_CgP1_validation(self):
-        """Test spacingCg_GP1_CgP1 parameter validation."""
-        base_airplane = geometry_fixtures.make_first_airplane_fixture()
-        wing_movements = [wing_movement_fixtures.make_static_wing_movement_fixture()]
-
-        # Test valid string values.
-        valid_spacings = [
-            ("sine", "sine", "sine"),
-            ("uniform", "uniform", "uniform"),
-            ("sine", "uniform", "sine"),
-        ]
-        for spacing in valid_spacings:
-            with self.subTest(spacing=spacing):
-                airplane_movement = ps.movements.airplane_movement.AirplaneMovement(
-                    base_airplane=base_airplane,
-                    wing_movements=wing_movements,
-                    spacingCg_GP1_CgP1=spacing,
-                )
-                self.assertEqual(airplane_movement.spacingCg_GP1_CgP1, spacing)
-
-        # Test invalid string raises error.
-        with self.assertRaises(ValueError):
-            ps.movements.airplane_movement.AirplaneMovement(
-                base_airplane=base_airplane,
-                wing_movements=wing_movements,
-                spacingCg_GP1_CgP1=("invalid", "sine", "sine"),
-            )
-
-    def test_phaseCg_GP1_CgP1_validation(self):
-        """Test phaseCg_GP1_CgP1 parameter validation."""
-        base_airplane = geometry_fixtures.make_first_airplane_fixture()
-        wing_movements = [wing_movement_fixtures.make_static_wing_movement_fixture()]
-
-        # Test valid phase values within range (-180.0, 180.0].
-        valid_phases = [
-            (0.0, 0.0, 0.0),
-            (90.0, 180.0, -90.0),
-            (179.9, 0.0, -179.9),
-        ]
-        for phase in valid_phases:
-            with self.subTest(phase=phase):
-                # Need non-zero amps for non-zero phases.
-                amp = tuple(1.0 if p != 0 else 0.0 for p in phase)
-                period = tuple(1.0 if p != 0 else 0.0 for p in phase)
-                airplane_movement = ps.movements.airplane_movement.AirplaneMovement(
-                    base_airplane=base_airplane,
-                    wing_movements=wing_movements,
-                    ampCg_GP1_CgP1=amp,
-                    periodCg_GP1_CgP1=period,
-                    phaseCg_GP1_CgP1=phase,
-                )
-                npt.assert_array_equal(airplane_movement.phaseCg_GP1_CgP1, phase)
-
-        # Test phase > 180.0 raises error.
-        with self.assertRaises(ValueError):
-            ps.movements.airplane_movement.AirplaneMovement(
-                base_airplane=base_airplane,
-                wing_movements=wing_movements,
-                ampCg_GP1_CgP1=(1.0, 1.0, 1.0),
-                periodCg_GP1_CgP1=(1.0, 1.0, 1.0),
-                phaseCg_GP1_CgP1=(180.1, 0.0, 0.0),
-            )
-
-        # Test phase <= -180.0 raises error.
-        with self.assertRaises(ValueError):
-            ps.movements.airplane_movement.AirplaneMovement(
-                base_airplane=base_airplane,
-                wing_movements=wing_movements,
-                ampCg_GP1_CgP1=(1.0, 1.0, 1.0),
-                periodCg_GP1_CgP1=(1.0, 1.0, 1.0),
-                phaseCg_GP1_CgP1=(-180.0, 0.0, 0.0),
-            )
-
-    def test_amp_period_relationship_Cg(self):
-        """Test that if ampCg_GP1_CgP1 element is 0, corresponding period must be 0."""
-        base_airplane = geometry_fixtures.make_first_airplane_fixture()
-        wing_movements = [wing_movement_fixtures.make_static_wing_movement_fixture()]
-
-        # Test amp=0 with period=0 works.
-        airplane_movement = ps.movements.airplane_movement.AirplaneMovement(
-            base_airplane=base_airplane,
-            wing_movements=wing_movements,
-            ampCg_GP1_CgP1=(0.0, 1.0, 0.0),
-            periodCg_GP1_CgP1=(0.0, 1.0, 0.0),
-        )
-        self.assertIsNotNone(airplane_movement)
-
-        # Test amp=0 with period!=0 raises error.
-        with self.assertRaises(ValueError):
-            ps.movements.airplane_movement.AirplaneMovement(
-                base_airplane=base_airplane,
-                wing_movements=wing_movements,
-                ampCg_GP1_CgP1=(0.0, 1.0, 0.0),
-                periodCg_GP1_CgP1=(1.0, 1.0, 0.0),
-            )
-
-    def test_amp_phase_relationship_Cg(self):
-        """Test that if ampCg_GP1_CgP1 element is 0, corresponding phase must be 0."""
-        base_airplane = geometry_fixtures.make_first_airplane_fixture()
-        wing_movements = [wing_movement_fixtures.make_static_wing_movement_fixture()]
-
-        # Test amp=0 with phase=0 works.
-        airplane_movement = ps.movements.airplane_movement.AirplaneMovement(
-            base_airplane=base_airplane,
-            wing_movements=wing_movements,
-            ampCg_GP1_CgP1=(0.0, 1.0, 0.0),
-            periodCg_GP1_CgP1=(0.0, 1.0, 0.0),
-            phaseCg_GP1_CgP1=(0.0, -90.0, 0.0),
-        )
-        self.assertIsNotNone(airplane_movement)
-
-        # Test amp=0 with phase!=0 raises error.
-        with self.assertRaises(ValueError):
-            ps.movements.airplane_movement.AirplaneMovement(
-                base_airplane=base_airplane,
-                wing_movements=wing_movements,
-                ampCg_GP1_CgP1=(0.0, 1.0, 0.0),
-                periodCg_GP1_CgP1=(0.0, 1.0, 0.0),
-                phaseCg_GP1_CgP1=(45.0, -90.0, 0.0),
-            )
-
-    def test_max_period_static_movement(self):
-        """Test that max_period returns 0.0 for static movement."""
-        airplane_movement = self.static_airplane_movement
-        self.assertEqual(airplane_movement.max_period, 0.0)
-
-    def test_max_period_Cg(self):
-        """Test that max_period returns correct period."""
-        airplane_movement = self.Cg_airplane_movement
-        # periodCg_GP1_CgP1 is (1.5, 1.5, 1.5), so max should be 1.5.
-        self.assertEqual(airplane_movement.max_period, 1.5)
-
-    def test_generate_airplanes_parameter_validation(self):
-        """Test that generate_airplanes validates num_steps and delta_time."""
-        airplane_movement = self.basic_airplane_movement
-
-        # Test invalid num_steps.
-        # noinspection PyTypeChecker
-        with self.assertRaises((ValueError, TypeError)):
-            airplane_movement.generate_airplanes(num_steps=0, delta_time=0.01)
-
-        # noinspection PyTypeChecker
-        with self.assertRaises((ValueError, TypeError)):
-            airplane_movement.generate_airplanes(num_steps=-1, delta_time=0.01)
-
-        with self.assertRaises(TypeError):
-            airplane_movement.generate_airplanes(num_steps="invalid", delta_time=0.01)
-
-        # Test invalid delta_time.
-        # noinspection PyTypeChecker
-        with self.assertRaises((ValueError, TypeError)):
-            airplane_movement.generate_airplanes(num_steps=10, delta_time=0.0)
-
-        # noinspection PyTypeChecker
-        with self.assertRaises((ValueError, TypeError)):
-            airplane_movement.generate_airplanes(num_steps=10, delta_time=-0.01)
-
-        with self.assertRaises(TypeError):
-            airplane_movement.generate_airplanes(num_steps=10, delta_time="invalid")
-
-    def test_generate_airplanes_returns_correct_length(self):
-        """Test that generate_airplanes returns list of correct length."""
-        airplane_movement = self.basic_airplane_movement
-
-        test_num_steps = [1, 5, 10, 50, 100]
-        for num_steps in test_num_steps:
-            with self.subTest(num_steps=num_steps):
-                airplanes = airplane_movement.generate_airplanes(
-                    num_steps=num_steps, delta_time=0.01
-                )
-                self.assertEqual(len(airplanes), num_steps)
-
-    def test_generate_airplanes_returns_correct_types(self):
-        """Test that generate_airplanes returns Airplanes."""
-        airplane_movement = self.basic_airplane_movement
-        airplanes = airplane_movement.generate_airplanes(num_steps=10, delta_time=0.01)
-
-        # Verify all elements are Airplanes.
-        for airplane in airplanes:
-            self.assertIsInstance(airplane, ps.geometry.airplane.Airplane)
-
-    def test_generate_airplanes_preserves_non_changing_attributes(self):
-        """Test that generate_airplanes preserves non-changing attributes."""
-        airplane_movement = self.basic_airplane_movement
-        base_airplane = airplane_movement.base_airplane
-
-        airplanes = airplane_movement.generate_airplanes(num_steps=10, delta_time=0.01)
-
-        # Check that non-changing attributes are preserved. Note: s_ref, c_ref,
-        # and b_ref are NOT included because they are calculated from the Wings,
-        # which change due to WingMovement or WingCrossSectionMovement.
-        for airplane in airplanes:
-            self.assertEqual(airplane.name, base_airplane.name)
-            self.assertEqual(airplane.weight, base_airplane.weight)
-
-    def test_generate_airplanes_static_movement(self):
-        """Test that static movement produces constant positions and angles."""
-        airplane_movement = self.static_airplane_movement
-        base_airplane = airplane_movement.base_airplane
-
-        airplanes = airplane_movement.generate_airplanes(num_steps=50, delta_time=0.01)
-
-        # All Airplanes should have same Cg_GP1_CgP1.
-        for airplane in airplanes:
-            npt.assert_array_equal(airplane.Cg_GP1_CgP1, base_airplane.Cg_GP1_CgP1)
-
-    def test_phase_offset_Cg(self):
-        """Test that phase shifts initial position correctly for Cg_GP1_CgP1."""
-        airplane_movement = self.phase_offset_Cg_airplane_movement
-        airplanes = airplane_movement.generate_airplanes(num_steps=100, delta_time=0.01)
-
-        # Extract positions (in Earth axes, relative to the simulation starting
-        # point).
-        x_positions = np.array([airplane.Cg_GP1_CgP1[0] for airplane in airplanes])
-        y_positions = np.array([airplane.Cg_GP1_CgP1[1] for airplane in airplanes])
-        z_positions = np.array([airplane.Cg_GP1_CgP1[2] for airplane in airplanes])
-
-        # Verify that phase offset causes non-zero initial values.
-        # With phase offsets, the first values should not all be at the base position.
-        self.assertFalse(np.allclose(x_positions[0], 0.0, atol=1e-10))
-        self.assertFalse(np.allclose(y_positions[0], 0.0, atol=1e-10))
-        self.assertFalse(np.allclose(z_positions[0], 0.0, atol=1e-10))
-
-    def test_single_dimension_movement_Cg(self):
-        """Test that only one dimension of Cg_GP1_CgP1 moves."""
-        airplane_movement = self.sine_spacing_Cg_airplane_movement
-        airplanes = airplane_movement.generate_airplanes(num_steps=50, delta_time=0.01)
-
-        # Extract positions (in Earth axes, relative to the simulation starting
-        # point).
-        x_positions = np.array([airplane.Cg_GP1_CgP1[0] for airplane in airplanes])
-        y_positions = np.array([airplane.Cg_GP1_CgP1[1] for airplane in airplanes])
-        z_positions = np.array([airplane.Cg_GP1_CgP1[2] for airplane in airplanes])
-
-        # Only x should vary, y and z should be constant.
-        self.assertFalse(np.allclose(x_positions, x_positions[0]))
-        npt.assert_array_equal(y_positions, y_positions[0])
-        npt.assert_array_equal(z_positions, z_positions[0])
-
-    def test_custom_spacing_function_Cg(self):
-        """Test that custom spacing function works for Cg_GP1_CgP1."""
-        airplane_movement = self.custom_spacing_Cg_airplane_movement
-        airplanes = airplane_movement.generate_airplanes(num_steps=100, delta_time=0.01)
-
-        # Extract x-positions (in Earth axes, relative to the simulation starting
-        # point).
-        x_positions = np.array([airplane.Cg_GP1_CgP1[0] for airplane in airplanes])
-
-        # Verify that values vary (not constant).
-        self.assertFalse(np.allclose(x_positions, x_positions[0]))
-
-        # Verify that values are within expected range.
-        # For custom_harmonic with amp=0.08, values should be in [-0.08, 0.08].
-        self.assertTrue(np.all(x_positions >= -0.09))
-        self.assertTrue(np.all(x_positions <= 0.09))
-
-    def test_custom_spacing_function_mixed_with_standard(self):
-        """Test that custom and standard spacing functions can be mixed."""
-        airplane_movement = self.mixed_custom_and_standard_spacing_airplane_movement
-        airplanes = airplane_movement.generate_airplanes(num_steps=100, delta_time=0.01)
-
-        # Verify that Airplanes are generated successfully.
-        self.assertEqual(len(airplanes), 100)
-        for airplane in airplanes:
-            self.assertIsInstance(airplane, ps.geometry.airplane.Airplane)
-
-
-class TestAirplaneMovementVariableGeometryOptimization(unittest.TestCase):
-    """This is a class with functions to test variable geometry optimization."""
-
-    @classmethod
-    def setUpClass(cls):
-        """Set up test fixtures once for all variable geometry optimization tests."""
-        cls.static_airplane_movement = (
-            airplane_movement_fixtures.make_static_airplane_movement_fixture()
-        )
-        cls.periodic_geometry_airplane_movement = (
-            airplane_movement_fixtures.make_periodic_geometry_airplane_movement_fixture()
-        )
-        cls.basic_airplane_movement = (
-            airplane_movement_fixtures.make_basic_airplane_movement_fixture()
-        )
-
-    def test_geometry_lcm_period_static(self):
-        """Test _geometry_lcm_period returns 0.0 for static geometry."""
-        result = self.static_airplane_movement._geometry_lcm_period()
-        self.assertEqual(result, 0.0)
-
-    def test_geometry_lcm_period_periodic(self):
-        """Test _geometry_lcm_period returns correct value for periodic geometry."""
-        # The periodic_geometry_airplane_movement has a 0.1s period.
-        result = self.periodic_geometry_airplane_movement._geometry_lcm_period()
-        self.assertAlmostEqual(result, 0.1, places=6)
-
-    def test_geometry_matches_identical_wings(self):
-        """Test _geometry_matches returns True for identical Wings."""
-        # Generate Wings for step 0.
-        wings = []
-        for wing_movement in self.static_airplane_movement.wing_movements:
-            step_0_wings = wing_movement.generate_wings(num_steps=1, delta_time=0.01)
-            wings.append(step_0_wings[0])
-
-        wings_array = np.array(wings)
-
-        # Identical Wings should match.
-        result = ps.movements.airplane_movement.AirplaneMovement._geometry_matches(
-            wings_step_a=wings_array,
-            wings_step_b=wings_array,
-            tolerance=1e-9,
-        )
-        self.assertTrue(result)
-
-    def test_geometry_matches_different_wings(self):
-        """Test _geometry_matches returns False for different Wings."""
-        # Generate Wings for two different time steps with movement.
-        airplane_movement = self.basic_airplane_movement
-        wings_step_0 = []
-        wings_step_5 = []
-
-        for wing_movement in airplane_movement.wing_movements:
-            all_wings = wing_movement.generate_wings(num_steps=10, delta_time=0.01)
-            wings_step_0.append(all_wings[0])
-            wings_step_5.append(all_wings[5])
-
-        wings_array_0 = np.array(wings_step_0)
-        wings_array_5 = np.array(wings_step_5)
-
-        # Different Wings should not match.
-        result = ps.movements.airplane_movement.AirplaneMovement._geometry_matches(
-            wings_step_a=wings_array_0,
-            wings_step_b=wings_array_5,
-            tolerance=1e-9,
-        )
-        self.assertFalse(result)
-
-    def test_geometry_matches_different_length(self):
-        """Test _geometry_matches returns False for different length Wing arrays."""
-        # Generate Wings.
-        wings = []
-        for wing_movement in self.static_airplane_movement.wing_movements:
-            step_0_wings = wing_movement.generate_wings(num_steps=1, delta_time=0.01)
-            wings.append(step_0_wings[0])
-
-        wings_array = np.array(wings)
-        # Create a shorter array.
-        wings_array_short = wings_array[:0]
-
-        result = ps.movements.airplane_movement.AirplaneMovement._geometry_matches(
-            wings_step_a=wings_array,
-            wings_step_b=wings_array_short,
-            tolerance=1e-9,
-        )
-        self.assertFalse(result)
-
-    def test_variable_geometry_optimization_applies(self):
-        """Test that variable geometry optimization applies for periodic motion."""
-        airplane_movement = self.periodic_geometry_airplane_movement
-
-        # Use delta_time = 0.01 and period = 0.1, so steps_per_period = 10.
-        # With 30 steps, we get 3 periods, so optimization should apply.
-        num_steps = 30
-        delta_time = 0.01
-
-        airplanes = airplane_movement.generate_airplanes(
-            num_steps=num_steps, delta_time=delta_time
-        )
-
-        # Verify correct number of Airplanes.
-        self.assertEqual(len(airplanes), num_steps)
-
-        # Verify all are Airplane instances.
-        for airplane in airplanes:
-            self.assertIsInstance(airplane, ps.geometry.airplane.Airplane)
-
-    def test_variable_geometry_periodicity(self):
-        """Test that variable geometry produces periodic results."""
-        airplane_movement = self.periodic_geometry_airplane_movement
-
-        # Use delta_time = 0.01 and period = 0.1, so steps_per_period = 10.
-        num_steps = 30
-        delta_time = 0.01
-        steps_per_period = 10
-
-        airplanes = airplane_movement.generate_airplanes(
-            num_steps=num_steps, delta_time=delta_time
-        )
-
-        # Check that geometry repeats at period boundaries.
-        # Compare step 0 to step 10 to step 20.
-        for period_num in range(1, 3):
-            base_step = 0
-            compare_step = period_num * steps_per_period
-
-            base_airplane = airplanes[base_step]
-            compare_airplane = airplanes[compare_step]
-
-            # Wings should have matching geometry.
-            for wing_id in range(len(base_airplane.wings)):
-                base_wing = base_airplane.wings[wing_id]
-                compare_wing = compare_airplane.wings[wing_id]
-
-                # Check Wing position.
-                npt.assert_allclose(
-                    base_wing.Ler_Gs_Cgs,
-                    compare_wing.Ler_Gs_Cgs,
-                    atol=1e-9,
-                    rtol=0.0,
-                )
-
-                # Check Wing angles.
-                npt.assert_allclose(
-                    base_wing.angles_Gs_to_Wn_ixyz,
-                    compare_wing.angles_Gs_to_Wn_ixyz,
-                    atol=1e-9,
-                    rtol=0.0,
-                )
-
-    def test_variable_geometry_Cg_updates(self):
-        """Test that Cg_GP1_CgP1 is updated correctly for deepcopied Airplanes."""
-        # Create an AirplaneMovement with both geometry motion and CG motion.
-        base_airplane = geometry_fixtures.make_first_airplane_fixture()
-        wing_movements = [
-            wing_movement_fixtures.make_periodic_geometry_wing_movement_fixture()
-        ]
-
-        airplane_movement = ps.movements.airplane_movement.AirplaneMovement(
-            base_airplane=base_airplane,
-            wing_movements=wing_movements,
-            ampCg_GP1_CgP1=(0.05, 0.0, 0.0),
-            periodCg_GP1_CgP1=(0.2, 0.0, 0.0),
-            spacingCg_GP1_CgP1=("sine", "sine", "sine"),
-            phaseCg_GP1_CgP1=(0.0, 0.0, 0.0),
-        )
-
-        num_steps = 30
-        delta_time = 0.01
-
-        airplanes = airplane_movement.generate_airplanes(
-            num_steps=num_steps, delta_time=delta_time
-        )
-
-        # Verify that Cg_GP1_CgP1 varies across steps (not all the same).
-        x_positions = [airplane.Cg_GP1_CgP1[0] for airplane in airplanes]
-        self.assertFalse(all(x == x_positions[0] for x in x_positions))
-
-    def test_fallback_when_period_not_aligned(self):
-        """Test that fallback to standard generation works when period not aligned."""
-        # Create an AirplaneMovement with a wing movement that has period = 1.0.
-        base_airplane = geometry_fixtures.make_first_airplane_fixture()
-        wing_movements = [
-            wing_movement_fixtures.make_sine_spacing_Ler_wing_movement_fixture()
-        ]
         airplane_movement = ps.movements.airplane_movement.AirplaneMovement(
             base_airplane=base_airplane,
             wing_movements=wing_movements,
         )
-
-        # Should still work (fallback to standard generation).
-        num_steps = 30
-        delta_time = 0.007  # Doesn't align with period = 1.0.
-
-        airplanes = airplane_movement.generate_airplanes(
-            num_steps=num_steps, delta_time=delta_time
-        )
-
-        self.assertEqual(len(airplanes), num_steps)
-
-    def test_no_optimization_when_single_period(self):
-        """Test that optimization doesn't apply when num_steps <= steps_per_period."""
-        airplane_movement = self.periodic_geometry_airplane_movement
-
-        # Period = 0.1, delta_time = 0.01, so steps_per_period = 10.
-        # Use num_steps = 10 (exactly one period). No benefit from optimization.
-        num_steps = 10
-        delta_time = 0.01
-
-        airplanes = airplane_movement.generate_airplanes(
-            num_steps=num_steps, delta_time=delta_time
-        )
-
-        # Should still work correctly.
-        self.assertEqual(len(airplanes), num_steps)
-        for airplane in airplanes:
-            self.assertIsInstance(airplane, ps.geometry.airplane.Airplane)
-
-
-class TestGeometryMatchesEdgeCases(unittest.TestCase):
-    """Tests for _geometry_matches edge cases and panel comparison code."""
-
-    @classmethod
-    def setUpClass(cls):
-        """Set up test fixtures once for all geometry matching tests."""
-        cls.static_airplane_movement = (
-            airplane_movement_fixtures.make_static_airplane_movement_fixture()
-        )
-        # Fixture with only angle movement (no position movement).
-        cls.angles_only_airplane_movement = (
-            airplane_movement_fixtures.make_angles_only_airplane_movement_fixture()
-        )
-
-    def test_geometry_matches_wing_angles_mismatch(self):
-        """Test _geometry_matches returns False when Wing angles don't match."""
-        # Use an AirplaneMovement with only angle movement (no position movement).
-        # This ensures Wing positions match but angles differ at different steps.
-        airplane_movement = self.angles_only_airplane_movement
-
-        # Generate Wings for two different time steps.
-        wings_step_0 = []
-        wings_step_5 = []
-
-        for wing_movement in airplane_movement.wing_movements:
-            all_wings = wing_movement.generate_wings(num_steps=10, delta_time=0.01)
-            wings_step_0.append(all_wings[0])
-            wings_step_5.append(all_wings[5])
-
-        wings_array_0 = np.array(wings_step_0)
-        wings_array_5 = np.array(wings_step_5)
-
-        # Verify positions are the same (angles-only movement doesn't change position).
-        for wing_0, wing_5 in zip(wings_array_0, wings_array_5):
-            npt.assert_allclose(
-                wing_0.Ler_Gs_Cgs, wing_5.Ler_Gs_Cgs, atol=1e-9, rtol=0.0
-            )
-
-        # Verify angles are different.
-        angles_differ = False
-        for wing_0, wing_5 in zip(wings_array_0, wings_array_5):
-            if not np.allclose(
-                wing_0.angles_Gs_to_Wn_ixyz,
-                wing_5.angles_Gs_to_Wn_ixyz,
-                atol=1e-9,
-                rtol=0.0,
-            ):
-                angles_differ = True
-                break
-        self.assertTrue(angles_differ, "Angles should differ between steps")
-
-        # _geometry_matches should return False due to angle mismatch.
-        result = ps.movements.airplane_movement.AirplaneMovement._geometry_matches(
-            wings_step_a=wings_array_0,
-            wings_step_b=wings_array_5,
-            tolerance=1e-9,
-        )
-        self.assertFalse(result)
-
-    def test_geometry_matches_panel_shape_mismatch(self):
-        """Test _geometry_matches returns False when Panel shapes don't match."""
-        # Create AirplaneMovements with different panel grid sizes.
-        airplane_movement_2 = (
-            airplane_movement_fixtures.make_2_chordwise_panels_airplane_movement_fixture()
-        )
-        airplane_movement_3 = (
-            airplane_movement_fixtures.make_3_chordwise_panels_airplane_movement_fixture()
-        )
-
-        # Generate Airplanes (which have meshed Wings with panels).
-        airplanes_2 = airplane_movement_2.generate_airplanes(
-            num_steps=1, delta_time=0.01
-        )
-        airplanes_3 = airplane_movement_3.generate_airplanes(
-            num_steps=1, delta_time=0.01
-        )
-
-        # Extract Wings from Airplanes.
-        wings_2 = airplanes_2[0].wings
-        wings_3 = airplanes_3[0].wings
-
-        wings_array_2 = np.array(wings_2)
-        wings_array_3 = np.array(wings_3)
-
-        # Verify Wings have panels.
-        self.assertIsNotNone(wings_2[0].panels)
-        self.assertIsNotNone(wings_3[0].panels)
-
-        # Verify panel shapes are different.
-        self.assertNotEqual(wings_2[0].panels.shape, wings_3[0].panels.shape)
-
-        # _geometry_matches should return False due to panel shape mismatch.
-        result = ps.movements.airplane_movement.AirplaneMovement._geometry_matches(
-            wings_step_a=wings_array_2,
-            wings_step_b=wings_array_3,
-            tolerance=1e-9,
-        )
-        self.assertFalse(result)
-
-    def _get_meshed_wings(self):
-        """Helper to get two copies of meshed Wings for panel corner tests."""
-        # Use static airplane movement to generate Airplanes with meshed Wings.
-        airplane_movement = self.static_airplane_movement
-
-        # Generate 2 time steps to get independent copies.
-        airplanes = airplane_movement.generate_airplanes(num_steps=2, delta_time=0.01)
-
-        # Extract Wings from each Airplane.
-        wings_a = list(airplanes[0].wings)
-        wings_b = list(airplanes[1].wings)
-
-        return wings_a, wings_b
-
-
-class TestVariableGeometryFallback(unittest.TestCase):
-    """Tests for the variable geometry fallback code path."""
-
-    def test_fallback_when_geometry_validation_fails(self):
-        """Test that fallback to standard generation works when geometry validation
-        fails.
-
-        This test uses unittest.mock to force _geometry_matches to return False,
-        triggering the fallback path at line 353.
-        """
-        from unittest.mock import patch
-
-        # Create a periodic AirplaneMovement.
-        airplane_movement = (
-            airplane_movement_fixtures.make_periodic_geometry_airplane_movement_fixture()
-        )
-
-        # Use parameters that would normally trigger optimization.
-        # Period = 0.1s, delta_time = 0.01s → steps_per_period = 10
-        # num_steps = 30 > steps_per_period, so optimization would apply.
-        num_steps = 30
-        delta_time = 0.01
-
-        # Mock _geometry_matches to return False, triggering fallback.
-        with patch.object(
+        self.assertIsInstance(
+            airplane_movement,
             ps.movements.airplane_movement.AirplaneMovement,
-            "_geometry_matches",
-            return_value=False,
-        ):
-            airplanes = airplane_movement.generate_airplanes(
-                num_steps=num_steps, delta_time=delta_time
-            )
-
-        # Verify correct number of Airplanes generated via fallback path.
-        self.assertEqual(len(airplanes), num_steps)
-
-        # Verify all are valid Airplane instances.
-        for airplane in airplanes:
-            self.assertIsInstance(airplane, ps.geometry.airplane.Airplane)
-
-
-class TestAirplaneMovementWingMovementsValidation(unittest.TestCase):
-    """Tests for wing_movements parameter validation in AirplaneMovement."""
-
-    def test_wing_movements_must_be_list(self):
-        """Test that wing_movements must be a list."""
-        base_airplane = geometry_fixtures.make_first_airplane_fixture()
-        wing_movement = wing_movement_fixtures.make_static_wing_movement_fixture()
-
-        # Test tuple raises TypeError.
-        with self.assertRaises(TypeError):
-            ps.movements.airplane_movement.AirplaneMovement(
-                base_airplane=base_airplane,
-                wing_movements=(wing_movement,),
-            )
-
-        # Test single WingMovement (not in list) raises TypeError.
-        with self.assertRaises(TypeError):
-            ps.movements.airplane_movement.AirplaneMovement(
-                base_airplane=base_airplane,
-                wing_movements=wing_movement,
-            )
-
-    def test_wing_movements_length_must_match_airplane_wings(self):
-        """Test that wing_movements length must match base_airplane.wings length."""
-        base_airplane = geometry_fixtures.make_first_airplane_fixture()
-        wing_movement = wing_movement_fixtures.make_static_wing_movement_fixture()
-
-        # base_airplane has 1 Wing, so 2 WingMovements should raise ValueError.
-        with self.assertRaises(ValueError):
-            ps.movements.airplane_movement.AirplaneMovement(
-                base_airplane=base_airplane,
-                wing_movements=[wing_movement, wing_movement],
-            )
-
-        # Empty list should raise ValueError.
-        with self.assertRaises(ValueError):
-            ps.movements.airplane_movement.AirplaneMovement(
-                base_airplane=base_airplane,
-                wing_movements=[],
-            )
-
-    def test_wing_movements_elements_must_be_wing_movements(self):
-        """Test that every element in wing_movements must be a WingMovement."""
-        base_airplane = geometry_fixtures.make_first_airplane_fixture()
-
-        # Test string raises TypeError.
-        with self.assertRaises(TypeError):
-            ps.movements.airplane_movement.AirplaneMovement(
-                base_airplane=base_airplane,
-                wing_movements=["not a wing movement"],
-            )
-
-        # Test None raises TypeError.
-        with self.assertRaises(TypeError):
-            ps.movements.airplane_movement.AirplaneMovement(
-                base_airplane=base_airplane,
-                wing_movements=[None],
-            )
-
-
-class TestAirplaneMovementImmutability(unittest.TestCase):
-    """Tests for AirplaneMovement attribute immutability."""
-
-    def setUp(self):
-        """Set up test fixtures for immutability tests."""
-        self.airplane_movement = (
-            airplane_movement_fixtures.make_basic_airplane_movement_fixture()
         )
 
-    def test_immutable_base_airplane_property(self):
-        """Test that base_airplane property is read only."""
-        new_airplane = geometry_fixtures.make_first_airplane_fixture()
-        with self.assertRaises(AttributeError):
-            self.airplane_movement.base_airplane = new_airplane
+    def test_rejects_core_wing_movement_children(self):
+        """Test that AirplaneMovement rejects CoreWingMovement instances."""
+        from tests.unit.fixtures import core_wing_movement_fixtures
 
-    def test_immutable_wing_movements_property(self):
-        """Test that wing_movements property is read only."""
-        new_wing_movements = [
-            wing_movement_fixtures.make_static_wing_movement_fixture()
+        base_airplane = geometry_fixtures.make_first_airplane_fixture()
+        wing_movements = [
+            core_wing_movement_fixtures.make_static_core_wing_movement_fixture()
         ]
-        with self.assertRaises(AttributeError):
-            self.airplane_movement.wing_movements = new_wing_movements
-
-    def test_wing_movements_returns_tuple(self):
-        """Test that wing_movements property returns a tuple (not a list)."""
-        wing_movements = self.airplane_movement.wing_movements
-        self.assertIsInstance(wing_movements, tuple)
-
-    def test_immutable_ampCg_GP1_CgP1_property(self):
-        """Test that ampCg_GP1_CgP1 property is read only."""
-        with self.assertRaises(AttributeError):
-            self.airplane_movement.ampCg_GP1_CgP1 = np.array([1.0, 2.0, 3.0])
-
-    def test_immutable_ampCg_GP1_CgP1_array_read_only(self):
-        """Test that ampCg_GP1_CgP1 array cannot be modified in place."""
-        with self.assertRaises(ValueError):
-            self.airplane_movement.ampCg_GP1_CgP1[0] = 999.0
-
-    def test_immutable_periodCg_GP1_CgP1_property(self):
-        """Test that periodCg_GP1_CgP1 property is read only."""
-        with self.assertRaises(AttributeError):
-            self.airplane_movement.periodCg_GP1_CgP1 = np.array([1.0, 2.0, 3.0])
-
-    def test_immutable_periodCg_GP1_CgP1_array_read_only(self):
-        """Test that periodCg_GP1_CgP1 array cannot be modified in place."""
-        with self.assertRaises(ValueError):
-            self.airplane_movement.periodCg_GP1_CgP1[0] = 999.0
-
-    def test_immutable_spacingCg_GP1_CgP1_property(self):
-        """Test that spacingCg_GP1_CgP1 property is read only."""
-        with self.assertRaises(AttributeError):
-            self.airplane_movement.spacingCg_GP1_CgP1 = (
-                "uniform",
-                "uniform",
-                "uniform",
+        with self.assertRaises(TypeError):
+            ps.movements.airplane_movement.AirplaneMovement(
+                base_airplane=base_airplane,
+                wing_movements=wing_movements,
             )
 
-    def test_spacingCg_GP1_CgP1_returns_tuple(self):
-        """Test that spacingCg_GP1_CgP1 property returns a tuple."""
-        spacing = self.airplane_movement.spacingCg_GP1_CgP1
-        self.assertIsInstance(spacing, tuple)
-
-    def test_immutable_phaseCg_GP1_CgP1_property(self):
-        """Test that phaseCg_GP1_CgP1 property is read only."""
-        with self.assertRaises(AttributeError):
-            self.airplane_movement.phaseCg_GP1_CgP1 = np.array([45.0, 45.0, 45.0])
-
-    def test_immutable_phaseCg_GP1_CgP1_array_read_only(self):
-        """Test that phaseCg_GP1_CgP1 array cannot be modified in place."""
-        with self.assertRaises(ValueError):
-            self.airplane_movement.phaseCg_GP1_CgP1[0] = 999.0
-
-
-class TestAirplaneMovementCaching(unittest.TestCase):
-    """Tests for AirplaneMovement caching behavior."""
-
-    def setUp(self):
-        """Set up test fixtures for caching tests."""
-        self.airplane_movement = (
-            airplane_movement_fixtures.make_multiple_periods_airplane_movement_fixture()
-        )
-
-    def test_all_periods_caching_returns_same_object(self):
-        """Test that repeated access to all_periods returns the same cached object."""
-        all_periods_1 = self.airplane_movement.all_periods
-        all_periods_2 = self.airplane_movement.all_periods
-        self.assertIs(all_periods_1, all_periods_2)
-
-    def test_max_period_caching_returns_same_value(self):
-        """Test that repeated access to max_period returns the same cached value."""
-        max_period_1 = self.airplane_movement.max_period
-        max_period_2 = self.airplane_movement.max_period
-        # Since floats are immutable, we check equality rather than identity.
-        self.assertEqual(max_period_1, max_period_2)
-
-
-class TestAirplaneMovementAllPeriods(unittest.TestCase):
-    """Tests for AirplaneMovement.all_periods property."""
-
-    @classmethod
-    def setUpClass(cls):
-        """Set up test fixtures once for all all_periods tests."""
-        cls.static_airplane_movement = (
-            airplane_movement_fixtures.make_static_airplane_movement_fixture()
-        )
-        cls.Cg_airplane_movement = (
-            airplane_movement_fixtures.make_Cg_airplane_movement_fixture()
-        )
-        cls.multiple_periods_airplane_movement = (
-            airplane_movement_fixtures.make_multiple_periods_airplane_movement_fixture()
-        )
-
-    def test_all_periods_static_movement(self):
-        """Test that all_periods returns empty tuple for static movement."""
-        airplane_movement = self.static_airplane_movement
-        self.assertEqual(airplane_movement.all_periods, ())
-
-    def test_all_periods_Cg_only_movement(self):
-        """Test that all_periods includes Cg periods."""
-        airplane_movement = self.Cg_airplane_movement
-        # periodCg_GP1_CgP1 is (1.5, 1.5, 1.5), all non zero.
-        # WingMovement is static, so no geometry periods.
-        # Should return tuple with three 1.5 values.
-        self.assertEqual(airplane_movement.all_periods, (1.5, 1.5, 1.5))
-
-    def test_all_periods_returns_tuple(self):
-        """Test that all_periods returns a tuple."""
-        airplane_movement = self.multiple_periods_airplane_movement
-        self.assertIsInstance(airplane_movement.all_periods, tuple)
-
-    def test_all_periods_includes_wing_movement_periods(self):
-        """Test that all_periods includes periods from WingMovements."""
-        airplane_movement = self.multiple_periods_airplane_movement
-        all_periods = airplane_movement.all_periods
-
-        # Should include periods from both Cg and WingMovements.
-        # Cg periods: (1.0, 2.0, 3.0).
-        # WingMovements contribute additional periods.
-        self.assertIn(1.0, all_periods)
-        self.assertIn(2.0, all_periods)
-        self.assertIn(3.0, all_periods)
-
-
-class TestAirplaneMovementDeepcopy(unittest.TestCase):
-    """Tests for AirplaneMovement.__deepcopy__ method."""
-
-    def setUp(self):
-        """Set up test fixtures for deepcopy tests."""
-        self.airplane_movement = (
+    def test_generate_airplanes_returns_airplanes(self):
+        """Test that generate_airplanes returns Airplanes when called through the
+        public class.
+        """
+        airplane_movement = (
             airplane_movement_fixtures.make_basic_airplane_movement_fixture()
         )
-
-    def test_deepcopy_returns_new_instance(self):
-        """Test that deepcopy returns a new AirplaneMovement instance."""
-        original = self.airplane_movement
-        copied = copy.deepcopy(original)
-
-        self.assertIsInstance(copied, ps.movements.airplane_movement.AirplaneMovement)
-        self.assertIsNot(original, copied)
-
-    def test_deepcopy_preserves_attribute_values(self):
-        """Test that deepcopy preserves all attribute values."""
-        original = self.airplane_movement
-        copied = copy.deepcopy(original)
-
-        # Check numpy array attributes.
-        npt.assert_array_equal(copied.ampCg_GP1_CgP1, original.ampCg_GP1_CgP1)
-        npt.assert_array_equal(copied.periodCg_GP1_CgP1, original.periodCg_GP1_CgP1)
-        npt.assert_array_equal(copied.phaseCg_GP1_CgP1, original.phaseCg_GP1_CgP1)
-
-        # Check tuple attributes.
-        self.assertEqual(copied.spacingCg_GP1_CgP1, original.spacingCg_GP1_CgP1)
-
-        # Check scalar derived properties.
-        self.assertEqual(copied.max_period, original.max_period)
-
-    def test_deepcopy_numpy_arrays_are_independent(self):
-        """Test that deepcopied numpy arrays are independent objects."""
-        original = self.airplane_movement
-        copied = copy.deepcopy(original)
-
-        # Verify arrays are different objects.
-        self.assertIsNot(copied.ampCg_GP1_CgP1, original.ampCg_GP1_CgP1)
-        self.assertIsNot(copied.periodCg_GP1_CgP1, original.periodCg_GP1_CgP1)
-        self.assertIsNot(copied.phaseCg_GP1_CgP1, original.phaseCg_GP1_CgP1)
-
-    def test_deepcopy_numpy_arrays_cannot_be_modified_in_place(self):
-        """Test that deepcopied numpy arrays raise ValueError on in place modification."""
-        original = self.airplane_movement
-        copied = copy.deepcopy(original)
-
-        # Verify that attempting to modify copied arrays raises ValueError.
-        with self.assertRaises(ValueError):
-            copied.ampCg_GP1_CgP1[0] = 999.0
-
-        with self.assertRaises(ValueError):
-            copied.periodCg_GP1_CgP1[0] = 999.0
-
-        with self.assertRaises(ValueError):
-            copied.phaseCg_GP1_CgP1[0] = 999.0
-
-    def test_deepcopy_base_airplane_is_independent(self):
-        """Test that deepcopied base_airplane is an independent object."""
-        original = self.airplane_movement
-        copied = copy.deepcopy(original)
-
-        # Verify base_airplane is a different object.
-        self.assertIsNot(copied.base_airplane, original.base_airplane)
-
-        # Verify attributes are equal.
-        self.assertEqual(copied.base_airplane.name, original.base_airplane.name)
-        self.assertEqual(copied.base_airplane.weight, original.base_airplane.weight)
-        npt.assert_array_equal(
-            copied.base_airplane.Cg_GP1_CgP1, original.base_airplane.Cg_GP1_CgP1
-        )
-
-    def test_deepcopy_wing_movements_are_independent(self):
-        """Test that deepcopied wing_movements are independent objects."""
-        original = self.airplane_movement
-        copied = copy.deepcopy(original)
-
-        # Verify wing_movements tuple is a different object.
-        self.assertIsNot(copied.wing_movements, original.wing_movements)
-
-        # Verify each WingMovement is a different object.
-        for original_wm, copied_wm in zip(
-            original.wing_movements, copied.wing_movements
-        ):
-            self.assertIsNot(copied_wm, original_wm)
-
-    def test_deepcopy_resets_caches_to_none(self):
-        """Test that deepcopy resets cached derived properties to None."""
-        original = self.airplane_movement
-
-        # Access cached properties to populate caches.
-        _ = original.all_periods
-        _ = original.max_period
-
-        # Verify original caches are populated.
-        self.assertIsNotNone(original._all_periods)
-        self.assertIsNotNone(original._max_period)
-
-        # Deepcopy the object.
-        copied = copy.deepcopy(original)
-
-        # Verify copied caches are reset to None.
-        self.assertIsNone(copied._all_periods)
-        self.assertIsNone(copied._max_period)
-
-    def test_deepcopy_cached_properties_can_be_recomputed(self):
-        """Test that cached properties work correctly after deepcopy."""
-        original = self.airplane_movement
-
-        # Get original cached values.
-        original_all_periods = original.all_periods
-        original_max_period = original.max_period
-
-        # Deepcopy the object.
-        copied = copy.deepcopy(original)
-
-        # Verify cached properties can be computed and match original.
-        self.assertEqual(copied.all_periods, original_all_periods)
-        self.assertEqual(copied.max_period, original_max_period)
-
-    def test_deepcopy_generate_airplanes_produces_same_results(self):
-        """Test that generate_airplanes produces same results after deepcopy."""
-        original = self.airplane_movement
-        copied = copy.deepcopy(original)
-
-        num_steps = 10
-        delta_time = 0.01
-
-        original_airplanes = original.generate_airplanes(
-            num_steps=num_steps, delta_time=delta_time
-        )
-        copied_airplanes = copied.generate_airplanes(
-            num_steps=num_steps, delta_time=delta_time
-        )
-
-        # Verify same number of Airplanes.
-        self.assertEqual(len(copied_airplanes), len(original_airplanes))
-
-        # Verify each Airplane has matching Cg_GP1_CgP1.
-        for original_ap, copied_ap in zip(original_airplanes, copied_airplanes):
-            npt.assert_array_equal(copied_ap.Cg_GP1_CgP1, original_ap.Cg_GP1_CgP1)
-            self.assertEqual(copied_ap.name, original_ap.name)
-            self.assertEqual(copied_ap.weight, original_ap.weight)
-
-    def test_deepcopy_handles_memo_correctly(self):
-        """Test that deepcopy handles the memo dict correctly for circular references."""
-        original = self.airplane_movement
-        memo = {}
-
-        # First deepcopy.
-        copied1 = copy.deepcopy(original, memo)
-
-        # Verify original is in memo.
-        self.assertIn(id(original), memo)
-        self.assertIs(memo[id(original)], copied1)
-
-        # Second deepcopy with same memo should return same object.
-        copied2 = copy.deepcopy(original, memo)
-        self.assertIs(copied1, copied2)
+        airplanes = airplane_movement.generate_airplanes(num_steps=5, delta_time=0.01)
+        self.assertEqual(len(airplanes), 5)
+        for airplane in airplanes:
+            self.assertIsInstance(
+                airplane,
+                ps.geometry.airplane.Airplane,
+            )
 
 
 if __name__ == "__main__":
